+++ oop-s2-3-mvc-78286/Controllers/RolesController.cs
@@ -3,13 +3,15 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using College.Domain.Models;
 using oop_s2_3_mvc_78286.Data;
+using Microsoft.AspNetCore.Authorization;
 
 namespace oop_s2_3_mvc_78286.Controllers
 {
+    // ONLY Administrators should ever see this controller
+    [Authorize(Roles = "Administrator")]
     public class RolesController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -22,115 +24,35 @@ namespace oop_s2_3_mvc_78286.Controllers
         // GET: Roles
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Roles.ToListAsync());
-        }
-
-        // GET: Roles/Details/5
-        public async Task<IActionResult> Details(string id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var role = await _context.Roles
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (role == null)
-            {
-                return NotFound();
-            }
-
-            return View(role);
-        }
-
-        // GET: Roles/Create
-        public IActionResult Create()
-        {
-            return View();
+            // Efficiency: NoTracking for a simple list
+            return View(await _context.Roles.AsNoTracking().ToListAsync());
         }
 
         // POST: Roles/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Name,NormalizedName,ConcurrencyStamp")] Role role)
+        public async Task<IActionResult> Create([Bind("Name")] Role role)
         {
             if (ModelState.IsValid)
             {
-                _context.Add(role);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(role);
-        }
-
-        // GET: Roles/Edit/5
-        public async Task<IActionResult> Edit(string id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var role = await _context.Roles.FindAsync(id);
-            if (role == null)
-            {
-                return NotFound();
-            }
-            return View(role);
-        }
-
-        // POST: Roles/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(string id, [Bind("Id,Name,NormalizedName,ConcurrencyStamp")] Role role)
-        {
-            if (id != role.Id)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
+                // RULE 1: Enforce Unique Role Names (Case-Insensitive)
+                bool exists = await _context.Roles.AnyAsync(r => r.Name.ToUpper() == role.Name.ToUpper());
+                if (exists)
                 {
-                    _context.Update(role);
+                    ModelState.AddModelError("Name", "This role already exists.");
+                }
+                else
+                {
+                    // RULE 2: Manually handle NormalizedName for Identity compatibility
+                    role.Id = Guid.NewGuid().ToString();
+                    role.NormalizedName = role.Name.ToUpper();
+                    role.ConcurrencyStamp = Guid.NewGuid().ToString();
+
+                    _context.Add(role);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!RoleExists(role.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
             }
-            return View(role);
-        }
-
-        // GET: Roles/Delete/5
-        public async Task<IActionResult> Delete(string id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var role = await _context.Roles
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (role == null)
-            {
-                return NotFound();
-            }
-
             return View(role);
         }
 
@@ -139,13 +61,29 @@ namespace oop_s2_3_mvc_78286.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
+            // RULE 3: Strict Referential Integrity
+            // Prevent deletion if any User is currently assigned to this Role
+            // (Assuming IdentityUserRole table exists)
+            bool isAssigned = await _context.UserRoles.AnyAsync(ur => ur.RoleId == id);
+
+            if (isAssigned)
+            {
+                return BadRequest("Cannot delete role. There are users currently assigned to this role.");
+            }
+
+            // RULE 4: Protect System-Critical Roles
             var role = await _context.Roles.FindAsync(id);
             if (role != null)
             {
+                if (role.Name == "Administrator")
+                {
+                    return BadRequest("The 'Administrator' role is protected and cannot be deleted.");
+                }
+
                 _context.Roles.Remove(role);
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
